refactor: simplify waitBeforeCapture logic in Storybook preview by replacing custom function with a fixed delay
